feat: expand types to allow differently typed arrays (#156)
* feat: expand types to allow differently typed arrays

With a comparator function, there's not a hard requirement for the two arrays to be compatible types. The comparator tells us whether they are "equal", and that's enough. This will maintain
the integrity between which type corresponds to what in the comparators, as well as the functions where appropriate.

This doesn't touch `getPatch` or `applyPatch`, because patching doesn’t make sense if the types don’t match.

* Add test cases for changed functions